feat: restructure sidebar components, implement new navigation elements, and enhance AWS settings UI
